Modified makefiles for Parallel Compilation
The makefiles have been modified to use the $(MAKE) variable instead of
the word make.  This allows the '-j N' flag to be passed to make which
tells it to spawn N processes and build in parallel. The following is
taken from the GNU make manual:

In order for make processes to communicate, the parent will pass
information to the child. Since this could result in problems if the
child process isn't actually a make, the parent will only do this if it
thinks the child is a make.
see http://www.gnu.org/software/make/manual/html_node/MAKE-Variable.html