Fixed SpCell::getNeighbors & SpCellList class. Added unit tests.
Added unit tests for SpCell and SpCellList classes based of the unit tests
for Cell and CellList classes in ddMd/neighbor.  Changed the way offsets
are calculated for finding neighboring cells based on a hybrid of the way
the grid and McMd::CellList calculate neighbors to take boundary conditions
into account.  Some unit tests need to be fixed and more need to be added
to test the SpCell::getNeighbors function.