Added PairEnergy Analyzer for Tools Namespace
This analyzer uses the cell list to loop over pairs of atoms and
calculate the pair energy.  Until an Interaction class is made for this
namespace, this only uses LJPair interactions.  Also it does not take
into account masks so it calculates energies as if there were a
MaskNone policy.