Added the ability to comment lines with a pound sign
This is useful for debugging / developing because you can just copy lines
and change them slightly while commenting out the old lines rather than
having to delete them when you might need them later.  Another option we
could do is to only consider lines beginning with a pound to be comments.

The function getNextLine is only called by the various Simulation objects
to read commands / config files.  It is also called by HoomdConfigReader
and LammpsDumpReader but those two file formats do not use the pound sign
so this shouldn't cause any problems.  If we still do not want this to
affect the reading of those file formats  I could implement a different
getNextLine function used only when the simulation objects read their
configs.